Phase 3: correctness fixes
- Widen fitness score u8 -> u16: max duplicates is 3*N*(N-1), which
  overflows u8 for N=16 (720) and N=25 (1800)
- Derive sub-box size with N.isqrt() instead of a hardcoded match, and
  use it in Display for Board/Row, fixing rendering for N != 9
- Validate GAParams: rate ranges and >= 2 survivors (previously an
  opaque divide-by-zero for e.g. --population 3 --selection-rate 0.5),
  fix stale doc comment; mirror validation as clap value_parsers so bad
  arguments are CLI errors, parsed before the terminal initializes
- Board::read: reject digits greater than N
- Replace hand-rolled TempFile test helper (fixed filenames, collision
  prone) with the tempfile crate; add 4x4 display and digit tests

+++ src/genetics/inner.rs
@@ -8,7 +8,7 @@ use rayon::vec::IntoIter;
 
 pub fn next_generation<const N: usize>(
     params: &super::GAParams,
-    population_scores: Vec<(Board<N>, u8)>,
+    population_scores: Vec<(Board<N>, u16)>,
 ) -> Vec<Board<N>> {
     make_parents(natural_selection(params, population_scores))
         .flat_map(|parents| make_children::<N>(params, parents))
@@ -17,7 +17,7 @@ pub fn next_generation<const N: usize>(
 
 fn natural_selection<const N: usize>(
     params: &super::GAParams,
-    mut population_scores: Vec<(Board<N>, u8)>,
+    mut population_scores: Vec<(Board<N>, u16)>,
 ) -> Vec<Board<N>> {
     population_scores.par_sort_by_key(|(_, score)| *score);
